adds styles file for unavailable projects Modal component
