--- src/components/search-projects-form/NoProjectsModal.tsx
@@ -1,50 +1,0 @@
-// Material UI
-import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
-import Typography from "@mui/material/Typography";
-import Modal from "@mui/material/Modal";
-// Context
-import { useContext } from "react";
-import { ErrorsContext } from "../../context/errors-context/ErrorsContext";
-
-const style = {
-    position: "absolute" as "absolute",
-    top: "50%",
-    left: "50%",
-    transform: "translate(-50%, -50%)",
-    width: 400,
-    bgcolor: "background.paper",
-    border: "2px solid #000",
-    boxShadow: 24,
-    p: 4,
-};
-
-const UnavailableProjectsModal = () => {
-    const errorContext = useContext(ErrorsContext);
-    const {
-        availableProjects,
-        handleOpen,
-        handleClose
-    } = errorContext;
-
-    return (
-        <div>
-            <Modal
-                open={ availableProjects }
-                onClose={ handleClose }
-                aria-labelledby="modal-modal-title"
-                aria-describedby="modal-modal-description"
-            >
-                <Box sx={ style }>
-                    <Typography id="modal-modal-title" variant="h6" component="h2">
-                        Text in a modal
-                    </Typography>
-                    <Button variant="outlined" onClick={ handleOpen }>
-                        Open modal
-                    </Button>
-                </Box>
-            </Modal>
-        </div>
-    );
-};
-export default UnavailableProjectsModal;
