+++ src/components/search-projects-form/unavailable-projects/styles.tsx
@@ -1,0 +1,14 @@
+export const modalContainer = {
+    position: "absolute" as "absolute",
+    top: "50%",
+    left: "50%",
+    transform: "translate(-50%, -50%)",
+    width: 400,
+    bgcolor: "background.paper",
+    border: "2px solid #000",
+    boxShadow: 24,
+    p: 4,
+    display: "flex",
+    flexDirection: "column",
+};
+export const closeModalButton = { width: "100px", alignSelf: "center" };
